fix weird circular dependancy
